Extract `DesignCTApp` to its own module
